pyproject.toml
[tool.poetry]
name = "hiddb"
version = "0.1.7"
description = "Python SDK for HIDDB"
keywords = [
    "HIDDB", 
    "hiddb", 
    "vector database", 
    "similarity search"
    ]
readme = "README.md"
license = "MIT"
homepage = "https://hiddb.com"
repository = "https://github.com/hiddb/hiddb-sdk-py.git"
include = [
    "LICENSE",
    ]
authors = ["Benjamin Bolbrinker <benjamin.bolbrinker@hiddb.com>"]

[tool.poetry.dependencies]
python = "^3.7"
PyJWT = "^2.3.0"
asyncio = "^3.4.3"
aiohttp = "^3.8.1"
requests = "^2.27.1"
urllib3 = "^1.26.9"
pytest-asyncio = "^0.18.3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
mypy = "^0.931"
black = "^22.1.0"
tox = "^3.24.5"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[pytest]
asyncio_mode="auto"
